# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/post_view_receipts:
    get:
      tags:
        - PostViewReceipts
      summary: List PostViewReceipts
      description: Get a paginated list of PostViewReceipts with optional filtering and sorting
      operationId: listPostViewReceipts
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostViewReceiptList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PostViewReceipts
      summary: Create PostViewReceipt
      description: Create a new PostViewReceipt
      operationId: createPostViewReceipt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostViewReceiptRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostViewReceipt'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/tag_categories:
    get:
      tags:
        - TagCategories
      summary: List TagCategories
      description: Get a paginated list of TagCategories with optional filtering and sorting
      operationId: listTagCategories
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagCategoryList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - TagCategories
      summary: Create TagCategory
      description: Create a new TagCategory
      operationId: createTagCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTagCategoryRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagCategory'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    PostViewReceipt:
      type: object
      required:
        - id
        - post_id
        - view_token
        - window_start
        - occurred_at
      properties:
        id:
          type: string
          format: uuid
        post_id:
          type: string
          format: uuid
        view_token:
          type: string
          maxLength: 64
        window_start:
          type: string
          format: date-time
        occurred_at:
          type: string
          format: date-time
...
    CreatePostViewReceiptRequest:
      type: object
      required:
        - post_id
        - view_token
        - window_start
        - occurred_at
      properties:
        post_id:
          type: string
          format: uuid
        view_token:
          type: string
          maxLength: 64
        window_start:
          type: string
          format: date-time
        occurred_at:
          type: string
          format: date-time
...
    PostViewReceiptList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PostViewReceipt'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    TagCategory:
      type: object
      required:
        - id
        - name
        - metadata
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 120
        metadata:
          type: object
...
    CreateTagCategoryRequest:
      type: object
      required:
        - name
        - metadata
      properties:
        name:
          type: string
          maxLength: 120
        metadata:
          type: object
...
    TagCategoryList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TagCategory'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []